refactor: change data_length type to uint16_t in build_cmd_frame method in LD2410S

+++ components/ld2410s/LD2410S.cpp
@@ -51,7 +51,7 @@ namespace esphome
             this->send_command(read_sn_cmd);
         }
 
-        CmdFrameT LD2410S::build_cmd_frame(uint16_t command, uint8_t* data, size_t data_length) {
+        CmdFrameT LD2410S::build_cmd_frame(uint16_t command, uint8_t* data, uint16_t data_length) {
             CmdFrameT cmd_frame = {
                 .header = CMD_FRAME_HEADER,
                 .command = command,
